refactor: Inventory, shop 패키지에 리팩토링
