+++ src/main/java/uos/uos25/shop/repository/InventoryRepository.java
@@ -1,4 +1,4 @@
-package uos.uos25.inventory.repository;
+package uos.uos25.shop.repository;
 
 import java.util.List;
 import java.util.Optional;
@@ -6,7 +6,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import uos.uos25.inventory.entity.Inventory;
+import uos.uos25.shop.entity.Inventory;
 
 @Repository
 public interface InventoryRepository extends JpaRepository<Inventory, Long> {
